+++ src/main/kotlin/com/sota/clone/copyopgg/models/LeagueSummoner.kt
@@ -1,0 +1,13 @@
+package com.sota.clone.copyopgg.models
+
+data class LeagueSummoner(
+    val summonerId: String,
+    val leagueId: String,
+    val leaguePoints: Int,
+    val wins: Int,
+    val loses: Int,
+    val veteran: Boolean,
+    val inactive: Boolean,
+    val freshBlood: Boolean,
+    val hotStreak: Boolean
+)
