+++ src/main/kotlin/com/sota/clone/copyopgg/models/Summoners.kt
@@ -20,40 +20,3 @@ data class SummonerBriefInfo(
     val leagueInfo: LeagueInfo?,
 )
 
-data class LeagueInfo(
-    val leagueId: String,
-    val tier: Tier,
-    val rank: Rank,
-    val leaguePoints: Int
-)
-
-data class LeagueSummoner(
-    val summonerId: String,
-    val leagueId: String,
-    val leaguePoints: Int,
-    val wins: Int,
-    val loses: Int,
-    val veteran: Boolean,
-    val inactive: Boolean,
-    val freshBlood: Boolean,
-    val hotStreak: Boolean
-)
-
-data class League(
-    val leagueId: String,
-    val tier: Tier,
-    val rank: Rank,
-    val queue: QueueType
-)
-
-enum class QueueType {
-    RANKED_SOLO_5x5, RANKED_FLEX_SR, RANKED_FLEX_TT
-}
-
-enum class Tier {
-    IRON, BRONZE, SILVER, GOLD, PLATINUM, DIAMOND
-}
-
-enum class Rank {
-    I, II, III, IV
-}
